added authorization functionality to post and comments

+++ controller/post.controller.js
@@ -18,21 +18,21 @@ class PostController {
 
   async getMyPosts(req, res) {
     try {
-      const {userId} = req.user.id;
+      const { userId } = req.user.id;
       const result = await pool.query(
         "SELECT * FROM posts WHERE author_id = $1 ORDER BY created_at DESC",
         [userId]
-      );  
+      );
       res.json(result.rows);
     } catch (err) {
       console.error(err);
       res.status(500).send("Error fetching user's posts");
-    } 
+    }
   }
 
   async getPosts(req, res) {
     try {
-      const {author} = req.query;
+      const { author } = req.query;
       let query = "SELECT * FROM posts WHERE status='published'";
 
       let params = [];
@@ -58,8 +58,8 @@ class PostController {
       const { id } = req.params;
 
       const postResult = await pool.query(
-        "SELECT p.id, p.title, p.content, p.status, p.created_at, u.username AS author " +
-        "FROM posts p JOIN users u ON p.author_id = u.id WHERE p.id = $1",
+        "SELECT p.id, p.title, p.content, p.status, p.created_at, p.author_id, u.username AS author " +
+          "FROM posts p JOIN users u ON p.author_id = u.id WHERE p.id = $1",
         [id]
       );
 
@@ -71,11 +71,22 @@ class PostController {
 
       const commentsResult = await pool.query(
         `SELECT c.id, c.content, c.created_at, u.username AS commenter ` +
-        `FROM comments c JOIN users u ON c.user_id = u.id WHERE c.post_id = $1 ORDER BY c.created_at ASC`,
+          `FROM comments c JOIN users u ON c.user_id = u.id WHERE c.post_id = $1 ORDER BY c.created_at ASC`,
         [id]
       );
 
-      post.coomments = commentsResult.rows;
+      post.comments = commentsResult.rows;
+
+      if (post.status !== "published") {
+        if (
+          !req.user ||
+          (req.user.role !== "admin" && req.user.id !== post.author_id)
+        ) {
+          return res
+            .status(403)
+            .send("Forbidden: Not authorized to view this post");
+        }
+      }
 
       res.json(post);
     } catch (err) {
@@ -87,14 +98,32 @@ class PostController {
   async updatePost(req, res) {
     try {
       const { id } = req.params;
+
+      const postResult = await pool.query(
+        `SELECT p.id , p.author_id
+        FROM posts p
+        WHERE p.id = $1`,
+        [id]
+      );
+
+      if (postResult.rows.length === 0) {
+        return res.status(404).send("Post not found");
+      }
+
+      const post = postResult.rows[0];
+
+      if (req.user.role !== "admin" && req.user.id !== post.author_id) {
+        return res
+          .status(403)
+          .send("Forbidden: You can only update your own posts");
+      }
+
       const { title, content, status } = req.body;
       const result = await pool.query(
         "UPDATE posts SET title = $1, content = $2, status = $3 WHERE id = $4 RETURNING *",
         [title, content, status, id]
       );
-      if (result.rows.length === 0) {
-        return res.status(404).send("Post not found");
-      }
+
       res.json(result.rows[0]);
     } catch (err) {
       console.error(err);
@@ -105,13 +134,30 @@ class PostController {
   async deletePost(req, res) {
     try {
       const { id } = req.params;
+
+      const postResult = await pool.query(
+        `SELECT p.id , p.author_id
+        FROM posts p
+        WHERE p.id = $1`,
+        [id]
+      );
+
+      if (postResult.rows.length === 0) {
+        return res.status(404).send("Post not found");
+      }
+
+      const post = postResult.rows[0];
+
+      if (req.user.role !== "admin" && req.user.id !== post.author_id) {
+        return res
+          .status(403)
+          .send("Forbidden: You can only delete your own posts");
+      }
       const result = await pool.query(
         "DELETE FROM posts WHERE id = $1 RETURNING *",
         [id]
       );
-      if (result.rows.length === 0) {
-        return res.status(404).send("Post not found");
-      }
+
       res.json({ message: "Post deleted", post: result.rows[0] });
     } catch (err) {
       console.error(err);
